novas aulas

+++ curso em video mundo01 python/exer035.py
@@ -1,7 +1,8 @@
-lado_1 = int(input('digite o primeiro lado do triangulo: '))
-lado_2 = int(input('digite o segundo lado do triangulo: '))
-lado_3 = int(input('digite o terceiro lado do triangulo: '))
+lado_1 = float(input('digite o primeiro lado do triangulo: '))
+lado_2 = float(input('digite o segundo lado do triangulo: '))
+lado_3 = float(input('digite o terceiro lado do triangulo: '))
 if lado_1 < lado_2 + lado_3 and lado_2 < lado_1 + lado_3 and lado_3 < lado_1 + lado_2:
     print('os lados formam um triangulo')
 else:
     print('os lados não formam um triangulo')
+
